Author: Roger Heflin Email: roger.heflin@gmail.com
Basic refcounter code, keeps track of the number of outstanding references,
adds on ref_hold, subtracts on ref_release.

Does the same thing with memory, on alloc adds the byte counts, on free
removes the bytes.

Adds code to several routines to print out the current counters at various
high level points in the code to allow memory leaks to be detected.

This code adds very little storage requirements and can likely be used on a
production system, and was tested on a production system with no apparent
issues.